requirements.txt
Jinja2
MarkupSafe
PyYAML
argparse
distribute
ecdsa
lxml
ncclient
paramiko
pexpect
pycrypto
requests
wsgiref
xmltodict
psycopg2
